+++ backend/src/main/kotlin/com/novatecgmbh/eventsourcing/axon/application/config/AxonConfig.kt
@@ -1,4 +1,4 @@
-package com.novatecgmbh.eventsourcing.axon.config
+package com.novatecgmbh.eventsourcing.axon.application.config
 
 import com.novatecgmbh.eventsourcing.axon.common.command.ExceptionWrappingHandlerInterceptor
 import org.axonframework.commandhandling.CommandBus
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-class AxonConfiguration {
+class AxonConfig {
 
   @Autowired
   fun commandBus(
