+++ backend/src/main/kotlin/com/novatecgmbh/eventsourcing/axon/application/security/CustomUserAuthenticationConverter.kt
@@ -1,4 +1,4 @@
-package com.novatecgmbh.eventsourcing.axon.security
+package com.novatecgmbh.eventsourcing.axon.application.security
 
 import org.springframework.core.convert.converter.Converter
 import org.springframework.security.authentication.AbstractAuthenticationToken
